Refactor to use core and prisma in one off scripts
